+++ broqer/op/__init__.py
@@ -8,10 +8,11 @@ from broqer.op.map_ import Map, build_map, build_map_factory
 from broqer.op.map_async import MapAsync, build_map_async, \
                                 build_map_async_factory, AsyncMode
 from broqer.op.bitwise import BitwiseCombineLatest, map_bit
+from broqer.op.cache import Cache
+from broqer.op.throttle import Throttle
 
 # utils
 from broqer.op.concat import Concat
-from broqer.op.throttle import Throttle
 
 # enable operator overloading
 from .py_operators import Str, Bool, Int, Float, Repr, Len, In, All, Any, \
@@ -23,5 +24,6 @@ __all__ = [
     'EvalFalse', 'build_map', 'build_map_factory', 'build_combine_latest',
     'build_filter', 'build_filter_factory', 'Concat', 'Str', 'Bool', 'Int',
     'Float', 'Repr', 'map_bit', 'build_map_async_factory',
-    'Len', 'In', 'All', 'Any', 'BitwiseAnd', 'BitwiseOr', 'Not', 'Throttle'
+    'Len', 'In', 'All', 'Any', 'BitwiseAnd', 'BitwiseOr', 'Not', 'Throttle',
+    'Cache'
 ]